commit to store current status

+++ src/commonroad_cpp/interfaces/commonroad/commonroad_factory_2018b.cpp
@@ -1,5 +1,8 @@
 //
-// Created by Sebastian Maierhofer on 08.11.20.
+// Created by Sebastian Maierhofer.
+// Technical University of Munich - Cyber-Physical Systems Group
+// Copyright (c) 2021 Sebastian Maierhofer - Technical University of Munich. All rights reserved.
+// Credits: BMW Car@TUM
 //
 
 #include "commonroad_factory_2018b.h"
@@ -71,7 +74,7 @@ CommonRoadFactory2018b::createLanelets(std::vector<std::shared_ptr<TrafficSign>>
                 }
                 if ((strcmp(child.name(), "speedLimit")) == 0) {
                     // laneletId + 4000 is the id of the virtual traffic sign element.
-                    int id = tempLaneletContainer[arrayIndex]->getId();
+                    size_t id = tempLaneletContainer[arrayIndex]->getId();
                     for (const auto &sign : trafficSigns) {
                         if (sign->getId() == id + 4000) {
                             tempLaneletContainer[arrayIndex]->addTrafficSign(sign);
